Fiducial Surface Angle Detection
A small program that gives the option to generate a fiducial marker and
also calculate the surface angle of the surface a maker is placed on
through the use of a webcam. This is a modified version of the samples
provided in the openCV aruco module. It has been tailored to work for
the Logitech C270 webcam.